all bugs are fixed
